make models a subpackage
to deal with absent imports more easily
