+++ geospaas_rest_api/models/processing_jobs.py
@@ -1,120 +1,17 @@
-"""Models for the GeoSPaaS REST API"""
+"""Jobs that rely on geospaas_processing"""
 from collections.abc import Sequence
-import functools
-import types
-try:
-    import geospaas_processing.tasks.core as tasks_core
-except ImportError as error:
-    tasks_core = error.name
-
-try:
-    import geospaas_processing.tasks.idf as tasks_idf
-except ImportError as error:
-    tasks_idf = error.name
-
-try:
-    import geospaas_processing.tasks.harvesting as tasks_harvesting
-except ImportError as error:
-    tasks_harvesting = error.name
-
-try:
-    import geospaas_processing.tasks.syntool as tasks_syntool
-except ImportError as error:
-    tasks_syntool = error.name
 
 import celery
 import dateutil.parser
-from celery.result import AsyncResult
-from django.db import models
+import geospaas_processing.tasks.core as tasks_core
+import geospaas_processing.tasks.idf as tasks_idf
+import geospaas_processing.tasks.harvesting as tasks_harvesting
+import geospaas_processing.tasks.syntool as tasks_syntool
 from rest_framework.exceptions import ValidationError
 
-
-def requires_func_decorator(*dependencies):
-    """Function decorator which raises an exception if a required
-    module is not available
-    """
-    def decorator(func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            missing_dependencies = []
-            for dependency in dependencies:
-                if not isinstance(dependency, types.ModuleType):
-                    missing_dependencies.append(dependency)
-            if missing_dependencies:
-                raise ImportError(f"{', '.join(d for d in missing_dependencies)} unavailable")
-            return func(*args, **kwargs)
-        return wrapper
-    return decorator
+from .base import Job
 
 
-def requires(*dependencies):
-    """Class decorator which applies requires_func_decorator() to the
-    __init__ method of a class
-    """
-    def decorator(cls):
-        cls.__init__ = requires_func_decorator(*dependencies)(cls.__init__)
-        return cls
-    return decorator
-
-
-class Job(models.Model):
-    """Base model that gives access to the status and result of
-    running one or more Celery tasks.
-    """
-    # Database fields
-    task_id = models.CharField(
-        unique=True, max_length=255,
-        help_text='ID of the first task in the job')
-    date_created = models.DateTimeField(
-        auto_now_add=True, db_index=True,
-        verbose_name='Creation DateTime',
-        help_text='Datetime: creation date of the job')
-
-    # Class attribute to be defined in child classes
-    signature = None
-
-    @classmethod
-    def get_signature(cls, parameters):
-        """Returns a Celery signature which will be executed when the
-        job is run. Can be one task or several organized using a
-        canvas.
-        See https://docs.celeryq.dev/en/stable/userguide/canvas.html
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def check_parameters(parameters):
-        """Checks that the parameters are valid for the current Job subclass"""
-        raise NotImplementedError
-
-    @staticmethod
-    def make_task_parameters(parameters):
-        """Returns the right task parameters from the request data"""
-        raise NotImplementedError
-
-    @classmethod
-    def run(cls, parameters):
-        """This method should be used to create jobs.
-        Should return a Job instance.
-        """
-        args, kwargs = cls.make_task_parameters(parameters)
-        result = cls.get_signature(parameters).delay(*args, **kwargs)
-        return cls(task_id=result.task_id)
-
-    def get_current_task_result(self):
-        """Get the AsyncResult of the currently running task"""
-        current_result = AsyncResult(self.task_id)
-        finished = False
-        while current_result.ready():
-            try:
-                current_result = current_result.children[0]
-            except IndexError:
-                finished = True
-                break
-        return current_result, finished
-
-
-@requires(tasks_core)
 class DownloadJob(Job):
     """
     Job which:
@@ -233,7 +130,6 @@ class ConvertJob(Job):  # pylint: disable=abstract-method
         return (((parameters['dataset_id'],),), {})
 
 
-@requires(tasks_syntool)
 class SyntoolCleanupJob(Job):
     """Job which cleans up ingested files older than a date"""
     class Meta:
@@ -253,12 +149,12 @@ class SyntoolCleanupJob(Job):
 
         try:
             dateutil.parser.parse(parameters['date'])
-        except KeyError:
-            raise ValidationError("The date parameter is mandatory")
+        except KeyError as error:
+            raise ValidationError("The date parameter is mandatory") from error
         except dateutil.parser.ParserError as error:
             raise ValidationError("'date' must be a valid date string") from error
 
-        if not isinstance(parameters.get('created', False), bool): # default is just a placeholder
+        if not isinstance(parameters.get('created', False), bool):  # default is just a placeholder
             raise ValidationError("'created' must be a boolean")
 
         return parameters
@@ -270,7 +166,6 @@ class SyntoolCleanupJob(Job):
             {'created': parameters.get('created', False)})
 
 
-@requires(tasks_harvesting)
 class HarvestJob(Job):
     """Job which harvests metadata into the database"""
     class Meta:
